+++ packages/headless/src/app/recommendation-app-reducers.ts
@@ -4,19 +4,19 @@ import {configurationReducer} from '../features/configuration/configuration-slic
 import {contextReducer} from '../features/context/context-slice';
 import {fieldsReducer} from '../features/fields/fields-slice';
 import {pipelineReducer} from '../features/pipeline/pipeline-slice';
+import {recommendationReducer} from '../features/recommendation/recommendation-slice';
 import {searchHubReducer} from '../features/search-hub/search-hub-slice';
-import {searchReducer} from '../features/search/search-slice';
 import {RecommendationAppState} from '../state/recommendation-app-state';
 
 /**
  * Map of reducers that make up the RecommendationAppState.
  */
-export const recommendationAppReducer: ReducersMapObject<RecommendationAppState> = {
+export const recommendationAppReducers: ReducersMapObject<RecommendationAppState> = {
   configuration: configurationReducer,
   advancedSearchQueries: advancedSearchQueriesReducer,
-  search: searchReducer,
   fields: fieldsReducer,
   context: contextReducer,
   pipeline: pipelineReducer,
   searchHub: searchHubReducer,
+  recommendation: recommendationReducer,
 };