+++ packages/quantic/force-app/main/default/lwc/searchBox/searchBox.js
@@ -1,4 +1,3 @@
-// @ts-check
 import {LightningElement, api, track} from 'lwc';
 import TributePath from '@salesforce/resourceUrl/tributejs';
 // @ts-ignore
@@ -8,6 +7,7 @@ import {initializeComponent} from 'c/initialization';
 export default class SearchBox extends LightningElement {
   /** @type {import("coveo").SearchBoxState} */
   @track state = {
+    // @ts-ignore TODO: Check SearchBoxState typing and integration with LWC/Quantic
     redirectTo: '',
     suggestions: [],
     value: '',
@@ -36,7 +36,9 @@ export default class SearchBox extends LightningElement {
   }
 
   disconnectedCallback() {
-    this.unsubscribe && this.unsubscribe();
+    if (this.unsubscribe) {
+      this.unsubscribe();
+    }
   }
 
   renderedCallback() {
@@ -79,10 +81,16 @@ export default class SearchBox extends LightningElement {
     this.tribute = new Tribute(tributeOptions);
     this.tribute.attach(input);
 
-    input.addEventListener('tribute-replaced', (e) => {
-      this.searchBox.updateText(e.detail.item.string);
-      this.searchBox.submit();
-    });
+    input.addEventListener(
+      'tribute-replaced',
+      /**
+       * @param {CustomEvent} e
+       */
+      (e) => {
+        this.searchBox.updateText(e.detail.item.string);
+        this.searchBox.submit();
+      }
+    );
 
     input.addEventListener('tribute-active-true', () => {
       combobox.classList.add('slds-is-open');