+++ packages/quantic/force-app/main/default/lwc/searchInterface/searchInterface.js
@@ -1,4 +1,3 @@
-// @ts-check
 import {LightningElement, api} from 'lwc';
 import HeadlessPath from '@salesforce/resourceUrl/coveoheadless';
 import AtomicPath from '@salesforce/resourceUrl/atomicutils';
@@ -54,7 +53,9 @@ export default class SearchInterface extends LightningElement {
       this.config = CoveoHeadless.HeadlessEngine.getSampleConfiguration();
     }
 
-    this.config && this.initEngine();
+    if (this.config) {
+      this.initEngine();
+    }
   }
 
   /**
@@ -77,13 +78,15 @@ export default class SearchInterface extends LightningElement {
       },
     };
 
-    this.dependenciesLoaded && this.initEngine();
+    if (this.dependenciesLoaded) {
+      this.initEngine();
+    }
   }
 
   initEngine() {
     this.engine = new CoveoHeadless.HeadlessEngine({
       configuration: this.config,
-      reducers: CoveoHeadless.searchPageReducers,
+      reducers: CoveoHeadless.searchAppReducers,
     });
 
     this.hangingComponents.forEach((component) =>