+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.29](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.28...v0.1.0-alpha.29) (2020-10-30)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.28](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.27...v0.1.0-alpha.28) (2020-10-29)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.27](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.26...v0.1.0-alpha.27) (2020-10-29)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.26](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.25...v0.1.0-alpha.26) (2020-10-29)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.25](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.24...v0.1.0-alpha.25) (2020-10-28)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.24](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.23...v0.1.0-alpha.24) (2020-10-28)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.23](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.22...v0.1.0-alpha.23) (2020-10-27)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.22](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.21...v0.1.0-alpha.22) (2020-10-27)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.21](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.20...v0.1.0-alpha.21) (2020-10-27)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.20](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.19...v0.1.0-alpha.20) (2020-10-26)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.1.0-alpha.19](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.18...v0.1.0-alpha.19) (2020-10-26)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 # [0.1.0-alpha.18](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.17...v0.1.0-alpha.18) (2020-10-23)
 
 **Note:** Version bump only for package @coveo/quantic